+++ backend/src/presentation/controllers/login/signup/signup-controller.spec.ts
@@ -1,13 +1,13 @@
-import { mockAccountModel } from '../../domain/test/mock-account';
-import { throwError } from '../../domain/test/test-helper';
-import { MissingParamError } from '../errors/missing-param-error';
-import { ValidationSpy } from '../test/mock-validation';
-import { Authentication } from '../../domain/usecases/account/authentication';
-import { EmailInUseError } from '../errors/email-in-use-error';
-import { ServerError } from '../errors/server-error';
-import { serverError, forbidden, ok, badRequest } from '../helpers/http/http-helper';
-import { HttpRequest } from '../protocols/http';
-import { AddAccountSpy, AuthenticationSpy } from '../test/mock-account';
+import { mockAccountModel } from '../../../../domain/test/mock-account';
+import { throwError } from '../../../../domain/test/test-helper';
+import { MissingParamError } from '../../../errors/missing-param-error';
+import { ValidationSpy } from '../../../test/mock-validation';
+import { Authentication } from '../../../../domain/usecases/account/authentication';
+import { EmailInUseError } from '../../../errors/email-in-use-error';
+import { ServerError } from '../../../errors/server-error';
+import { serverError, forbidden, ok, badRequest } from '../../../helpers/http/http-helper';
+import { HttpRequest } from '../../../protocols/http';
+import { AddAccountSpy, AuthenticationSpy } from '../../../test/mock-account';
 import { SignUpController } from './signup-controller';
 
 const mockRequest = (): HttpRequest => ({
